Punt this annoying test to flakey
